updated room booking logic
suppose dates are 3 to 10 and  their  is a booking for day 4-7 and 5-6. The hotel has only 2 rooms so anyone can't book room for day 4-7. But according to your logic  if i want to book room for day  3 to 8 it will be successful ( Because their is no any booking for date less than 4 and greater than 7) while all the middle days are booked. So it is also important to check that the middle days are booked or not. Above logic will solve this problem.